+++ NEWS.md
@@ -1,5 +1,7 @@
 # pakret (development version)
 
+* pakret now creates the `.bib` files used to write package references if they don't exist (#48).
+
 * pakret now preserves the case of cited packages and single-letter names (such as R and C) in reference titles (#43).
 
 * pakret now works when called after the package [conflicted](https://conflicted.r-lib.org) (#30).